+++ app/graphql/types/product_type.rb
@@ -15,6 +15,6 @@ module Types
     field :unit_size, String, null: true
     field :documents, [DocumentDataType], null: true
     field :portion_size, Integer, null: true
-    field :marked_up_price, Integer, null: true
+    field :marked_up_price, Float, null: true
   end
 end
